- fix issues with ids when upload screenshots; - update target to 35 and compile to 34; - add  automation test that will send screenshot;

+++ crowdin/src/main/java/com/crowdin/platform/data/remote/api/CrowdinApi.kt
@@ -32,7 +32,7 @@ internal interface CrowdinApi {
     @POST("api/v2/projects/{projectId}/screenshots/{screenshotId}/tags")
     fun createTag(
         @Path("projectId") projectId: String,
-        @Path("screenshotId") screenshotId: Int,
+        @Path("screenshotId") screenshotId: Long,
         @Body tags: MutableList<TagData>,
     ): Call<ResponseBody>
